pyproject.toml
[project]
name = "AI-Driven-Coaching-Feedback-system"
version = "0.1.0"
description = "Instructional coaching feedback system"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "Your Name"}]
dependencies = [
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.36.0",
    "rich>=13.7.0",
    "jinja2>=3.1.4",
    "strawberry-graphql>=0.219.2",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "PyJWT[crypto]>=2.9.0",
    "requests>=2.32.3",
    "boto3>=1.34.0",
    "faster-whisper>=0.10.0",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "pyttsx3>=2.90",
    "pydub>=0.25.1",
    "jsonschema>=4.23.0"
]

[project.optional-dependencies]
mqtt = ["paho-mqtt>=1.6.1"]
kafka = ["kafka-python>=2.0.2"]
dev = ["pytest>=7.4.0", "ruff>=0.5.0", "black>=24.4.0", "mypy>=1.10.0"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
markers = [
    "cloud: cloud-dependent tests",
]
